% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEL_functions.R
\name{BEL_leroux_new}
\alias{BEL_leroux_new}
\title{BEL spatial function for BYM and Leroux}
\usage{
BEL_leroux_new(
  y,
  x,
  n,
  p,
  var,
  rho,
  niter,
  beta_init,
  psi_init,
  tau_init,
  R,
  wi,
  sd_psi,
  sd_beta,
  sd_tau
)
}
\arguments{
\item{y}{observed response from data}

\item{x}{design matrix including a column of 1 for intercept and the other from dataset}

\item{n}{number of observations/small areas in the dataset}

\item{p}{number of regression paramters or number of columns in design matrix x}

\item{var}{residual variance}

\item{rho}{fixed value for spatial dependence rho=1 gives BYM rho=0 gives non spatial model}

\item{niter}{number of iterations}

\item{beta_init}{initial values for beta}

\item{psi_init}{initial values of spatial random effect psi}

\item{tau_init}{initial value of precision paramter tau}

\item{R}{neighbourhood matrix from spatial data}

\item{wi}{current EL weights}

\item{sd_psi}{value of sd for proposal distribution for psi}

\item{sd_beta}{value of sd for proposal distribution for beta}

\item{sd_tau}{value of sd for proposal distribution for tau}
}
\description{
BEL spatial function for BYM and Leroux
}
