% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{beta_giv_tau}
\alias{beta_giv_tau}
\title{Supporting function 1 to compute Conditional Posterior for tau}
\usage{
beta_giv_tau(b, b_mean, tau, g = 10)
}
\arguments{
\item{b}{current value of beta}

\item{b_mean}{mean of the proposal distribution}

\item{tau}{current value of precision parameter}

\item{g}{value of zellnor's g prior taken as 10}
}
\description{
Supporting function 1 to compute Conditional Posterior for tau
}
