% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bel_leroux_gradient.R
\name{dg}
\alias{dg}
\title{gradient function for estimating equations of beta}
\usage{
dg(tet, x, var, y)
}
\arguments{
\item{tet}{initial value of regression coefficients beta}

\item{x}{design matrix}

\item{var}{residual variance from the fitted model and data}

\item{y}{observed response variable}
}
\description{
gradient function for estimating equations of beta
}
