% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bel_leroux_gradient.R
\name{g1}
\alias{g1}
\title{Estimating equation for GEL estimates of beta}
\usage{
g1(tet, x, var, y)
}
\arguments{
\item{tet}{initial value of regression coefficients beta}

\item{x}{design matrix}

\item{var}{residual variance from the fitted model and data}

\item{y}{observed response variable}
}
\description{
Estimating equation for GEL estimates of beta
}
