% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{M_create}
\alias{M_create}
\title{For Porter (2015) creation of matrix M}
\usage{
M_create(y, x, B)
}
\arguments{
\item{y}{observed data for response variable}

\item{x}{design matrix}

\item{B}{adjacency matrix}
}
\description{
For Porter (2015) creation of matrix M
}
