% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{MBM_create}
\alias{MBM_create}
\title{For Porter (2015) creation of matrix M}
\usage{
MBM_create(M, B, Bplus)
}
\arguments{
\item{M}{n × q matrix with the columns being the eigenvectors corresponding to the q largest nonzero eigenvalues of the matrix PBP,P be the latent process space and B be the adjacency matrix}

\item{B}{adjacency matrix}

\item{Bplus}{diagional matrix stating the number of neghbours of each small area}
}
\description{
For Porter (2015) creation of matrix M
}
