% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bel_leroux_gradient.R
\name{mele}
\alias{mele}
\title{generalised Empirical likelihood estimation for of beta}
\usage{
mele(x, tet, y, var)
}
\arguments{
\item{x}{design matrix}

\item{tet}{Initial value of regression coefficients}

\item{y}{observed response}

\item{var}{residual variance}
}
\description{
generalised Empirical likelihood estimation for of beta
}
