% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{psi_giv_tau}
\alias{psi_giv_tau}
\title{Supporting function 2 to compute Conditional Posterior for tau}
\usage{
psi_giv_tau(psi, tau, D)
}
\arguments{
\item{psi}{current value of psi}

\item{tau}{current value of precision parameter}

\item{D}{variance covariance matrix of spatial random effects}
}
\description{
Supporting function 2 to compute Conditional Posterior for tau
}
