% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{target_beta}
\alias{target_beta}
\title{Conditional Posterior for Beta}
\usage{
target_beta(b, w, b_mean, g, tau)
}
\arguments{
\item{b}{current value of beta}

\item{w}{current EL weights}

\item{b_mean}{mean of the proposal distribution}

\item{g}{value of zellnor's g prior taken as 10}

\item{tau}{current value of precision  parameter}
}
\description{
Conditional Posterior for Beta
}
