% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{target_psi}
\alias{target_psi}
\title{Conditional Posterior for psi}
\usage{
target_psi(w, psi, D, tau)
}
\arguments{
\item{w}{current EL weights}

\item{psi}{current value of psi}

\item{D}{variance covariance matrix of spatial random effects}

\item{tau}{current value of precision parameter}
}
\description{
Conditional Posterior for psi
}
