% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{target_tau_porter}
\alias{target_tau_porter}
\title{For Porter (2015) conditional posterior of tau}
\usage{
target_tau_porter(
  tau,
  beta,
  beta_mean,
  alpha_1,
  alpha_2,
  psi,
  MBM,
  g,
  p,
  q = q
)
}
\arguments{
\item{tau}{current value of precision parameter}

\item{beta}{current value of beta}

\item{beta_mean}{mean of the proposal distribution}

\item{alpha_1}{shape parameter for tau inverse gamma prior.}

\item{alpha_2}{scale parameter for tau inverse gamma prior.}

\item{psi}{current value of psi}

\item{MBM}{MBM M(B_plus-B)M, where B_plus is a diagonal matrix having number of neighbours for each area in its diagonal, a positive definite matrix, variance covariance for spatial random effects}

\item{g}{value of zellnor's g prior taken as 10}

\item{p}{number of fixed effect parameters}

\item{q}{number of positive eigenvalues of PBP (Porter 2015)}
}
\description{
For Porter (2015) conditional posterior of tau
}
