% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{target_yq}
\alias{target_yq}
\title{For Porter (2015) conditional posterior of yq}
\usage{
target_yq(a, w, MBM, tau)
}
\arguments{
\item{a}{current/proposed value of yq, dimension reduced psi, spatial random effects}

\item{w}{EL weights}

\item{MBM}{MBM M(B_plus-B)M, where B_plus is a diagonal matrix having number of neighbours for each area in its diagonal, a positive definite matrix, variance covariance for spatial random effects}

\item{tau}{current value of precision parameter}
}
\description{
For Porter (2015) conditional posterior of yq
}
