% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{tau_prior}
\alias{tau_prior}
\title{Supporting function 3 to compute Conditional Posterior for tau}
\usage{
tau_prior(tau, alpha_1, alpha_2)
}
\arguments{
\item{tau}{current value of precision parameter}

\item{alpha_1}{scale parameter of inverse gamma distribution}

\item{alpha_2}{shape parameter of inverse gamma distribution}
}
\description{
Supporting function 3 to compute Conditional Posterior for tau
}
