% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{yq_giv_tau}
\alias{yq_giv_tau}
\title{For Porter (2015) supporting function for conditional posterior of tau}
\usage{
yq_giv_tau(a, tau, MBM)
}
\arguments{
\item{a}{current/proposed value of yq, dimension reduced psi, spatial random effects}

\item{tau}{current value of precision parameter}

\item{MBM}{M(B_plus-B)M, where B_plus is a diagonal matrix having number of neighbours for each area in its diagonal, a positive definite matrix,variance covariance for spatial random effects}
}
\description{
For Porter (2015) supporting function for conditional posterior of tau
}
